% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_allRNAmod_metrics}
\alias{add_allRNAmod_metrics}
\title{Add all metrics for STORMseq v0.0.2}
\usage{
add_allRNAmod_metrics(STORM)
}
\arguments{
\item{STORM}{}
}
\value{

}
\description{
Add all metrics for STORMseq v0.0.2
}
