% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_default_metrics}
\alias{add_default_metrics}
\title{Add all default metrics for STORMseq}
\usage{
add_default_metrics(STORM, nCores = 1)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{nCores}{}
}
\description{
Add all default metrics for STORMseq
}
