% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_default_metrics_v1}
\alias{add_default_metrics_v1}
\title{Add all default metrics for STORMseq v1}
\usage{
add_default_metrics_v1(STORM)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated as the STORMseq library treatments were expanded.
This functions applied only for the library treatments of lib 499.
}
