% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_default_metrics_v2}
\alias{add_default_metrics_v2}
\title{Add all default metrics for STORMseq v0.0.2}
\usage{
add_default_metrics_v2(STORM)
}
\arguments{
\item{STORM}{}
}
\value{

}
\description{
Add all default metrics for STORMseq v0.0.2
}
