% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_DRD}
\alias{add_DRD}
\title{Add deletion rate difference}
\usage{
add_DRD(
  STORM,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minReads = 30,
  nCores = 1
)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{group_A}{character. Name of group to be compared as found in STORM$META$group}

\item{group_B}{character. Name of group to be compared as found in STORM$META$group}

\item{newColName}{character. Name of calculated metric to be stored in STORM$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}

\item{minReads}{numeric. Minimum number of nucleotides read in position to
compute the output value.}
}
\description{
Calculated the deletion rate to nucleotide reads for each position and
subtracts it for the rate in the control group.
}
