% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_ERlog2FCh1bpDS}
\alias{add_ERlog2FCh1bpDS}
\title{Add End rate log2-Fold-Change 1bp downstream}
\usage{
add_ERlog2FCh1bpDS(
  STORM,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50,
  nCores = 1
)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{group_A}{character. Name of group to be compared as found in STORM$META$group}

\item{group_B}{character. Name of group to be compared as found in STORM$META$group}

\item{newColName}{character. Name of calculated metric to be stored in STORM$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}

\item{minCov}{integer. Minimal coverage in position for metric to be calculated}
}
\description{
Calculates the log2-Fold-Change of the end rate ratio 1bp downstream
of group_A over that of group_B per set of samples in a STORM object.
}
