% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_MRDns}
\alias{add_MRDns}
\title{Add Nucleotide Specific Misincorporation Rate Difference}
\usage{
add_MRDns(STORM, group_A, group_B, refNuc, misNuc, minReads = 30, nCores = 1)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{group_A}{character. Name of group to be compared as found in STORM$META$group}

\item{group_B}{character. Name of group to be compared as found in STORM$META$group}

\item{refNuc}{character. Reference nucleotide}

\item{misNuc}{character. Misincorporated nucleotide}

\item{minReads}{numeric. Minimum number of nucleotides read in position to
compute the output value.}

\item{newColName}{character. Name of calculated metric to be stored in STORM$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}
}
\description{
Add Nucleotide Specific Misincorporation Rate Difference
}
