% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_scoreC3p}
\alias{add_scoreC3p}
\alias{storm_add_scoreC3p}
\title{Add ScoreC accounting for read-stops (3prime)}
\usage{
add_scoreC3p(
  STORM,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minReads = 30,
  flankSize = 6,
  nCores = 1
)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{group_A}{character. Name of group to be compared as found in STORM$META$group}

\item{newColName}{character. Name of calculated metric to be stored in STORM$RES,
assigned by default based on group_A and group_B}

\item{onNucs}{character. Nucleotide(s) in which the metric will be calculated}

\item{minReads}{integer. Minimum number of reads needed at site neighborhood
to calculate the score.}

\item{flankSize}{integer. Length of window to each side of position to
consider for metric calculation}

\item{minMedCov}{integer. Minimal median coverage of window around position,
delimited by flankSize}
}
\description{
Add ScoreC accounting for read-stops (3prime)
}
