% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_SR}
\alias{add_SR}
\title{Add read-start to coverage ratio}
\usage{
add_SR(
  STORM,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50
)
}
\arguments{
\item{minCov}{}
}
\value{

}
\description{
Add read-start to coverage ratio
}
