% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_SRlog2FCh1bpDS}
\alias{add_SRlog2FCh1bpDS}
\title{Add Start rate log2-Fold-Change 1bp down-stream}
\usage{
add_SRlog2FCh1bpDS(
  STORM,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50
)
}
\arguments{
\item{minCov}{}
}
\value{

}
\description{
Calculates the log2-Fold-Change of the start ratio difference 1bp down-stream
of group_A over that of group_B per set of samples in a STORM object.
}
