% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-clustering-functions.R
\name{add_table_to_workbook}
\alias{add_table_to_workbook}
\title{Add annotated pathways table to Excel workbook.}
\usage{
add_table_to_workbook(annot_table, wb, sheet_name, tab_color)
}
\arguments{
\item{annot_table}{A data.frame or data.table with the annotated pathways data.}

\item{wb}{A workbook object created by \code{openxlsx::createWorkbook()}.}

\item{sheet_name}{A character string specifying the worksheet name.}

\item{tab_color}{A character string representing the hex color code for the worksheet tab.}
}
\value{
This function does not return a value; it modifies the workbook object in place.
}
\description{
This function adds an annotated pathways table to an Excel workbook with appropriate formatting and styles.
}
