% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-functions.R
\name{add_worksheet}
\alias{add_worksheet}
\title{Add Excel worksheet.}
\usage{
add_worksheet(
  wb,
  sheet,
  table,
  header = h1(),
  row1 = r1(),
  row2 = r2(),
  sc = 1,
  sr = 1,
  cw = 30,
  rh = 20,
  fc = F,
  fr = T,
  fac = NULL,
  far = NULL,
  tab = "#1A1B41",
  ...
)
}
\arguments{
\item{wb}{\code{openxlsx} Workbook object.}

\item{sheet}{Worksheet name.}

\item{table}{Table to add to worksheet.}

\item{header}{Header \code{openxlsx} style object generated by \code{brainstorm::h1()}.}

\item{row1}{Primary row \code{openxlsx} style object generated by \code{brainstorm::r1()}.}

\item{row2}{Secondary row \code{openxlsx} style object generated by \code{brainstorm::r2()}.}

\item{sc}{Starting column to write to.}

\item{sr}{Starting row to write to.}

\item{cw}{Vector of column widths or \code{"auto"} for automatic sizing.}

\item{rh}{Vector of row heights or \code{"auto"} for automatic sizing.}

\item{fc}{If \code{TRUE}, freezes first column. To freeze more than the first column, set as \code{FALSE} and specify \code{fac}.}

\item{fr}{If \code{TRUE}, freezes first row. To freeze more than the first row, set as \code{FALSE} and specify \code{far}.}

\item{fac}{First active column, where \code{fac = 2} is equivalent to \code{fc = TRUE}.}

\item{far}{First active row, where \code{far = 2} is equivalent to \code{fr = TRUE}.}

\item{tab}{Tab color of worksheet.}

\item{...}{Other arguments to be passed to \code{openxlsx::addWorksheet()}.}
}
\description{
Combine \code{brainstorm} functions to add Excel worksheet. For finer control, functions can be called individually. Depends on the \code{openxlsx} package.
}
\author{
Ayush Noori
}
