% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_STORM.R
\name{addKnownRNAmods}
\alias{addKnownRNAmods}
\title{Add known RNA modifications nucleotide identity}
\usage{
addKnownRNAmods(STORM, RNAmods)
}
\arguments{
\item{STORM}{STORM object}

\item{RNAmods}{data.frame with two columns "pos" for the positions in the
same format as the tables inside the STORM object as STORM$RES, a combination
in the form of "gene:txcoor"; And "nuc" for the identity of the nucleotide
on said position.}
}
\description{
Add the nucleotide identity based on a table with columns "pos" and "nuc
establishing the position and nucleotide identity
Useful to annotate already known RNA modifications. The function will add
a column called "nuc" in the "RES" and "CALLS" tables.
}
