% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-clustering-functions.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{Annotate pathway clusters for manual review.}
\usage{
annotate_clusters(dt)
}
\arguments{
\item{dt}{A \code{data.table} containing pathways data with a column named \code{Cluster}.}
}
\value{
A \code{data.frame} with inserted header rows and computed averages for each cluster.
}
\description{
This function inserts header rows into a pathways data table for manual annotation
and computes average values for each cluster.
}
