% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-functions.R
\name{apply_r1}
\alias{apply_r1}
\title{Apply Excel primary row style.}
\usage{
apply_r1(
  wb,
  sheet,
  table = NULL,
  style = r1(),
  rows = 2:(nrow(table) + 1) \%>\% .[which(.\%\%2 == 0)],
  cols = 1:ncol(table),
  grid = T,
  stack = F
)
}
\arguments{
\item{wb}{\code{openxlsx} Workbook object.}

\item{sheet}{Worksheet name.}

\item{table}{Table for specified worksheet. If both \code{rows} and \code{cols} are specified, then this argument is optional.}

\item{style}{\code{openxlsx} style object generated by \code{brainstorm::r1()}.}

\item{rows}{Rows to apply style to, optional if  \code{table} is specified..}

\item{cols}{Columns to apply style to, optional if  \code{table} is specified..}

\item{grid}{See \code{gridExpand} argument of \code{openxlsx::addStyle()}.}

\item{stack}{See \code{stack} argument of \code{openxlsx::addStyle()}.}
}
\description{
Apply Excel primary row style to odd rows. Depends on the \code{openxlsx} package. All arguments have default values and are optional to allow parameters to be modified.
}
\author{
Ayush Noori
}
