% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_STORM.R
\name{bam2TxDT}
\alias{bam2TxDT}
\title{BAM to txDataTable}
\usage{
bam2TxDT(
  BAMfile,
  geneAnnot,
  genome,
  dtType,
  paired = TRUE,
  outDir,
  remL = 1000,
  minR = 50,
  nCores = 2,
  ySize = 1e+05,
  verb = TRUE
)
}
\arguments{
\item{verb}{}
}
\description{
From BAM file to count data in a table
}
