% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_STORM.R
\name{bam2txDT_META}
\alias{bam2txDT_META}
\title{BAM 2 txDT from META}
\usage{
bam2txDT_META(
  META,
  nCores,
  geneAnnot,
  genome,
  outDir,
  remL = 1000,
  minR = 50,
  force = FALSE,
  verb = TRUE
)
}
\arguments{
\item{minR}{}
}
\description{
BAM 2 txDT from META
}
