% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-functions.R
\name{base_theme}
\alias{base_theme}
\title{Return base theme.}
\usage{
base_theme()
}
\value{
\code{ggplot2} theme.
}
\description{
Return base theme from the \code{ggplot2} package.
}
\author{
Ayush Noori
}
