% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{bisGenome}
\alias{bisGenome}
\title{Make bisulphite genome}
\usage{
bisGenome(fastaGenome, outFile = "auto")
}
\arguments{
\item{fastaGenome}{character. Path to genome fasta file}

\item{outFile}{character. Name of output file. Auto adds}
}
\description{
Make bisulphite genome
}
