% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brain_sentrix}
\alias{brain_sentrix}
\title{Brain Sentrix Table}
\format{
\code{data.frame}
}
\source{
\url{https://github.com/LieberInstitute/SPEAQeasy-example}
}
\usage{
brain_sentrix
}
\description{
Genotype sample sheet for 42 samples from 39 unique brains
}
\keyword{datasets}
