% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainSTORM-package.R
\docType{package}
\name{brainSTORM-package}
\alias{brainSTORM}
\alias{brainSTORM-package}
\title{brainSTORM: brainSTORM. A package to process STORM-seq based data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package to process STORM-seq based data.
}
\author{
\strong{Maintainer}: Miguel Angel Garcia-Campos \email{angel.campos@ciencias.unam.mx} (\href{https://orcid.org/0000-0001-5223-8174}{ORCID}) (https://angelcampos.github.io/)

}
\keyword{internal}
