% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{called}
\alias{called}
\title{Example RNA snps (imputed Genotypes)}
\format{
\code{CollapsedVCF}
}
\source{
\url{https://github.com/LieberInstitute/SPEAQeasy-example}
}
\usage{
called
}
\description{
Imputed snps from RNA data, 665 snps for 42 samples
}
\keyword{datasets}
