% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_STORM.R
\name{check_whichRNAmods}
\alias{check_whichRNAmods}
\title{Check RNAmods}
\usage{
check_whichRNAmods(STORM)
}
\arguments{
\item{STORM}{}
}
\value{
character
}
\description{
Check which RNAmods are possible to predict based on RNAmod metrics present
in STORM object
}
