% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{fastq_nucFreq}
\alias{fastq_nucFreq}
\title{Nucleotide frequency report}
\usage{
fastq_nucFreq(META, nCores, firstN = 10000, pairedEnd = TRUE)
}
\arguments{
\item{pairedEnd}{}
}
\description{
Nucleotide frequency report
}
