% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_called.R
\name{filter_called}
\alias{filter_called}
\title{Filter the Called snp array}
\usage{
filter_called(
  called,
  DP_min = 5,
  DP_max = 80,
  ref_call = 1,
  alt_call = 1,
  overlaps = 1,
  min_VDB = 0.1
)
}
\arguments{
\item{called}{CollapsedVCF containing filtered snp info for rna data}

\item{DP_min}{TODO}

\item{DP_max}{TODO}

\item{ref_call}{TODO}

\item{alt_call}{TODO}

\item{overlaps}{TODO}

\item{min_VDB}{TODO}
}
\value{
Filtered snp array
}
\description{
Filter the Called snp array
}
\examples{
dim(snpsCalled_VCF)
snpsCalled_filter <- filter_called(snpsCalled_VCF)
dim(snpsCalled_filter)
}
