% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genotyped}
\alias{genotyped}
\title{TODO genotyped}
\format{
A CollapsedVCF object.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/genotyped.R}
}
\usage{
genotyped
}
\description{
TODO Description.
}
\keyword{datasets}
