% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{gg_lce}
\alias{gg_lce}
\title{Library complexity extrapolation plot}
\usage{
gg_lce(META, tab_name, speciesName = "")
}
\arguments{
\item{speciesName}{}
}
\description{
Library complexity extrapolation plot
}
