% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouper.R
\name{grouper}
\alias{grouper}
\title{grouper}
\usage{
grouper(corTable, cutoff = 0.59, s1 = "row_sample", s2 = "col_sample")
}
\arguments{
\item{corTable}{Table of paiwise correlation between samples}

\item{cutoff}{corrlation threshold for samples to be grouped together}

\item{s1}{Name of first column that contains sample names}

\item{s2}{Name of second column that contains sample names}

\item{samples}{list of samples (BrNum and/or RNum) to subset cor table by}
}
\value{
List of groups found in the cor table
}
\description{
Create groups of sample from a cor table with correlation >= cutoff
}
\examples{
cor_test <- make_corLong(snps1 = test_snps, BrainTable1 = test_BrTable, ID_col1 = "Sample")
g <- grouper(cor_test)
}
