% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-functions.R
\name{h1}
\alias{h1}
\title{Create Excel header style.}
\usage{
h1(
  fill = "#1A1B41",
  color = "#FFFFFF",
  font = "Arial Black",
  size = 10,
  h = "center",
  v = "center",
  style = "Bold",
  border = "Bottom",
  bstyle = "thick",
  ...
)
}
\arguments{
\item{fill}{Header foreground fill.}

\item{color}{Header font color.}

\item{font}{Header font name.}

\item{size}{Header font size.}

\item{h}{Horizontal alignment of header contents.}

\item{v}{Vertical alignment of header contents.}

\item{style}{Header text styling.}

\item{border}{Header cell border.}

\item{bstyle}{Header border line style.}

\item{...}{Other arguments to be passed to \code{openxlsx::createStyle()}.}
}
\value{
\code{openxlsx} style object.
}
\description{
Create Excel header style. Depends on the \code{openxlsx} package. All arguments have default values and are optional to allow parameters to be modified.
}
\author{
Ayush Noori
}
