% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hlp_assign_scores}
\alias{hlp_assign_scores}
\title{Assign scores to respective RNAmods in STORM$CALLS}
\usage{
hlp_assign_scores(STORM, RNAmod, scores, onNucs = c("A", "C", "G", "T"))
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}.}

\item{RNAmod}{character. Name of RNAmod to assign scores to.}

\item{scores}{character. Names of scores to be assigned to RNAmod.}

\item{onNucs}{character. Vector of base nucleotides to include in scores, all
nucleotides by default, i.e. c("A", "C", "G", "T").}
}
\description{
Assign scores to respective RNAmods in STORM$CALLS
}
