% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hlp_start_CALLS}
\alias{hlp_start_CALLS}
\title{Initialize STORM$CALLS}
\usage{
hlp_start_CALLS(STORM)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}
}
\description{
Initialize STORM$CALLS
}
