% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_corLong.R
\name{make_corLong}
\alias{make_corLong}
\title{Create Long Correlation Tables}
\usage{
make_corLong(
  snps1,
  snps2 = snps1,
  BrainTable1,
  BrainTable2 = BrainTable1,
  ID_col1 = "ID",
  ID_col2 = ID_col1
)
}
\arguments{
\item{snps1}{Table of snps, RNA or DNA,}

\item{snps2}{Table of snps, RNA or DNA, if NA snps1 will be compared to its self}

\item{BrainTable1}{Table containing ID info for samples in snps1}

\item{BrainTable2}{Table containing ID info for samples in snps2}

\item{ID_col1}{Name of column containing Samples names in BrainTable1}

\item{ID_col2}{Name of column containing Samples names in BrainTable2}
}
\value{
Data frame with correlation between samples in snps1 (or between snps1 and snps2)
and ID info from BrainTables (at least BrNums for samples)
}
\description{
Compute pairwise correlations between one or two snp arrays, then convert to long format
}
\examples{
x <- make_corLong(snps1 = test_snps, BrainTable1 = test_BrTable, BrainTable2 = test_BrTable, ID_col1 = "Sample", ID_col2 = "Sample")

snpsGeno <- make_snpsGeno(genotyped)
snpsGeno[1:5, 1:5]
brain_sentrix <-
    read.csv(system.file(
        "extdata",
        "brain_sentrix.csv",
        package = "brainstorm",
        mustWork = TRUE
    ))
corLongtest <- make_corLong(snpsGeno, BrainTable1 = brain_sentrix)

corLongtest <- make_corlong
}
