% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snpsGeno.R
\name{make_snpsGeno}
\alias{make_snpsGeno}
\title{Title}
\usage{
make_snpsGeno(genotyped)
}
\arguments{
\item{genotype}{object from variantAnnotation::readVcf}
}
\value{

}
\description{
Title
}
\examples{
genotyped <- readVcf("Data/LIBD_Brain_merged_topmed_RNASeq_annotated_variants_043020.vcf", "hg38")
snpsGeno <- make_snpsGeno(genotyped)
jaffelab::corner(snpsGeno)
head(colnames(snpsGeno))
}
