% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snpsRNA.R
\name{make_snpsRNA}
\alias{make_snpsRNA}
\title{make_snpsRNA}
\usage{
make_snpsRNA(genotyped, called_filtered)
}
\arguments{
\item{genotyped}{CollapsedVCF containing snp info for genotype data}

\item{called_filtered}{CollapsedVCF containing filtered snp info for rna data (generated by filtered_called)}
}
\value{
Named list containing snpsGeno2 and snpsCalled, both representing snp
arrays as integers.
}
\description{
Prep rna snps and create DNA & RNA snps (snpsGeno2) and RNA snps (snpsCalled)
}
\examples{

called_filter <- filter_called(called)
snpsRNA <- make_snpsRNA(genotyped, called_filtered)
jaffelab::corner(snpsRNA$snpsGeno2)
jaffelab::corner(snpsRNA$snpsCalled)
}
