% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_STORM.R
\name{mkBedFromFastaTxOme}
\alias{mkBedFromFastaTxOme}
\title{Make BED from FASTA Transcriptome}
\usage{
mkBedFromFastaTxOme(fastaTxOme, outFile = "auto")
}
\arguments{
\item{outFile}{}
}
\value{

}
\description{
Make GeneAnnotation for Transcriptome
}
