% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_STORM.R
\name{mkSTARgenome}
\alias{mkSTARgenome}
\title{Make STAR genome}
\usage{
mkSTARgenome(
  fastaGenome,
  bedAnnotation = NULL,
  outDir = NULL,
  nCores = 2,
  maxReadLength = 101
)
}
\arguments{
\item{maxReadLength}{}
}
\value{

}
\description{
Wrapper for STAR genome generation
}
