% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_example}
\alias{pd_example}
\title{Example Phenotype Table}
\format{
\code{data.frame}
}
\source{
\url{https://github.com/LieberInstitute/SPEAQeasy-example}
}
\usage{
pd_example
}
\description{
Phenotype data for 42 samples
}
\keyword{datasets}
