% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-functions.R
\name{r1}
\alias{r1}
\title{Create Excel primary row style.}
\usage{
r1(
  fill = "#FFFFFF",
  color = "#363635",
  font = "Arial",
  size = 10,
  h = "center",
  v = "center",
  border = "TopBottomLeftRight",
  ...
)
}
\arguments{
\item{fill}{Primary row foreground fill.}

\item{color}{Primary row font color.}

\item{font}{Primary row font name.}

\item{size}{Primary row font size.}

\item{h}{Horizontal alignment of primary row contents.}

\item{v}{Vertical alignment of primary row contents.}

\item{border}{Primary row cell border.}

\item{...}{Other arguments to be passed to \code{openxlsx::createStyle()}.}
}
\value{
\code{openxlsx} style object.
}
\description{
Create Excel primary row style. Depends on the \code{openxlsx} package. All arguments have default values and are optional to allow parameters to be modified.
}
\author{
Ayush Noori
}
