% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-functions.R
\name{r2}
\alias{r2}
\title{Create Excel secondary row style.}
\usage{
r2(
  fill = "#F6F4F4",
  color = "#363635",
  font = "Arial",
  size = 10,
  h = "center",
  v = "center",
  border = "TopBottomLeftRight",
  ...
)
}
\arguments{
\item{fill}{Secondary row foreground fill.}

\item{color}{Secondary row font color.}

\item{font}{Secondary row font name.}

\item{size}{Secondary row font size.}

\item{h}{Horizontal alignment of secondary row contents.}

\item{v}{Vertical alignment of secondary row contents.}

\item{border}{Secondary row cell border.}

\item{...}{Other arguments to be passed to \code{openxlsx::createStyle()}.}
}
\value{
\code{openxlsx} style object.
}
\description{
Create Excel secondary row style. Depends on the \code{openxlsx} package. All arguments have default values and are optional to allow parameters to be modified.
}
\author{
Ayush Noori
}
