% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{restoreGenCoors}
\alias{restoreGenCoors}
\title{Restore txDT Genomic Coordinate System}
\usage{
restoreGenCoors(txDT, geneAnnot, nCores = 1)
}
\arguments{
\item{txDT}{data.table}

\item{geneAnnot}{GenomicRanges Gene annotation loaded via tx_load_bed()}

\item{nCores}{integer}
}
\value{
data.table
}
\description{
Restore txDT Genomic Coordinate System
}
