% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAmod_metrics}
\alias{RNAmod_metrics}
\title{RNAmod metrics table}
\format{
data.frame
}
\usage{
RNAmod_metrics
}
\description{
Table with all RNAmod metrics used in brainSTORM with their respective RNAmod,
groups, RTase, and function used to generate it.
}
\keyword{datasets}
