% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNAmods_vec}
\alias{RNAmods_vec}
\title{STORM base coordinate columns}
\format{
character
}
\usage{
RNAmods_vec
}
\description{
Used internally by brainSTORM but may be useful
}
\keyword{datasets}
