% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rRNAmods_Sc_Taoka}
\alias{rRNAmods_Sc_Taoka}
\title{Yeast rRNA modifications reference}
\format{
data.frame
}
\usage{
rRNAmods_Sc_Taoka
}
\description{
rRNA modifications in yeast according to Taoka et al.,
2016 (Nucleic acids research)
}
\keyword{datasets}
