% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-functions.R
\name{show_table}
\alias{show_table}
\title{Show Markdown table.}
\usage{
show_table(table)
}
\arguments{
\item{table}{Table to display.}
}
\value{
`kable` table object.
}
\description{
Read table in Markdown document. Depends on `knitr` and `kableExtra` packages.
}
\author{
Ayush Noori
}
