% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpsGeno_example}
\alias{snpsGeno_example}
\title{Example RNA Genotype Data}
\format{
\code{CollapsedVCF}
}
\source{
\url{https://github.com/LieberInstitute/SPEAQeasy-example}
}
\usage{
snpsGeno_example
}
\description{
TOPMed Imputed genotype data (746 snps) for the samples in pd_example
}
\keyword{datasets}
