% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpsGeno_VCF}
\alias{snpsGeno_VCF}
\title{Example Genotype Data}
\format{
\code{CollapsedVCF}
}
\source{
\url{https://github.com/LieberInstitute/SPEAQeasy-example}
}
\usage{
snpsGeno_VCF
}
\description{
TOPMed Genotype data (746 snps) for the 42 samples in brain_sentrix
}
\keyword{datasets}
