% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{STARalign_META}
\alias{STARalign_META}
\title{Align using STAR from META}
\usage{
STARalign_META(
  META,
  STARGenome,
  STARGenomebis,
  pairedEnd = TRUE,
  outDir,
  nCores,
  tmpDir,
  logSumm,
  alignIntronMax = 0,
  otherSTARparams = "",
  force = FALSE
)
}
\arguments{
\item{alignIntronMax}{numeric. maximum intron size, if 0, max intron size will be determined
by default as (2ˆwinBinNbits)*winAnchorDistNbins (See STAR manual for more info)}

\item{otherSTARparams}{character. Additional parameters not covered by the arguments
of this function can be added here, separated by spaces.}

\item{force}{logical. Set to TRUE to realign existing BAM files.}
}
\description{
Align using STAR from META
}
