% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{storm_assignScores}
\alias{storm_assignScores}
\title{Assigning default Scores to tables by sets of samples}
\usage{
storm_assignScores(STORM)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}
}
\description{
Assigning default Scores to tables by sets of samples
}
