% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storm_baseCols}
\alias{storm_baseCols}
\title{STORM base columns}
\format{
character
}
\usage{
storm_baseCols
}
\description{
Used internally by brainSTORM but may be useful
}
\keyword{datasets}
