% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{storm_baseCoorCols}
\alias{storm_baseCoorCols}
\title{STORM base coordinate columns}
\format{
character
}
\usage{
storm_baseCoorCols
}
\description{
Used internally by brainSTORM but may be useful
}
\keyword{datasets}
