% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{storm_evalPreds}
\alias{storm_evalPreds}
\title{Evaluate predictions}
\usage{
storm_evalPreds(STORM, nucBias = TRUE, roundDig = 4)
}
\arguments{
\item{STORM}{list. STORM object}

\item{nucBias}{logical. Nucleotide bias: only cases in the base nucleotide
of the respective RNAmod are considered.}

\item{roundDig}{numeric. Number of digits to round the resulting specificity,
sensitivity, FDR, and query rate values.}
}
\description{
Evaluate predictions using known modified nucleotides identity
}
