% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_STORM.R
\name{storm_META}
\alias{storm_META}
\title{Create STORM META - Experiment design matrix}
\usage{
storm_META(
  fileNames,
  varsList = brainSTORM::vList,
  groupVars = c("libTreat", "RTase"),
  setVars = c("libNum", "organism", "bioTreat", "replicate"),
  idVars = c("libNum", "organism", "RTase", "libTreat", "bioTreat", "replicate"),
  outDir,
  bisTreats = c("m5C", "RBSseqHeatMg")
)
}
\arguments{
\item{fileNames}{character. Full path to read1-FASTQ files, file names must
contain the "_R1" suffix.}

\item{varsList}{list. A list of char vectors that define a sample and their
belonging to a group, and set of samples. Each element of the list must have
a name that specifies it.}

\item{groupVars}{character. The names of variables used to define a group
of samples. By default the library treatment ("libTreat") and reverse
transcriptase ("RTase") in that order. Changing this only allows for flexibility
on the definition of comparable samples. \strong{Should not be changed unless having}
\strong{a deep understanding of the brainSTORM package and require to develop new}
** functionality**}

\item{setVars}{character. Vector of variables from varsList, which combination
define the belonging of a group of FASTQ read files to the same biological
sample}

\item{idVars}{character. Vector of variable names from varsList, which
combination define the identity of each FASTQ read file. None combination
of variables should result in a duplicated id.}

\item{outDir}{character. Path to directory where all processed files
(BAM and RDS) will be written to.}

\item{bisTreats}{character. Vector with names of library treatments that
require to be aligned with a bisulfite version}
}
\description{
STORM META constructor from RDS file names
}
