% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_storm.R
\name{storm_metricsBoxPlot_byNuc}
\alias{storm_metricsBoxPlot_byNuc}
\title{Boxplots of $RES calculated metrics grouping by modified nucleotides}
\usage{
storm_metricsBoxPlot_byNuc(STORM, metrics = NULL, title = "")
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{metrics}{character. Selected metrics to plot. By default plots all metrics}

\item{title}{character. Title to be shown on top of plot}
}
\description{
Boxplots of $RES calculated metrics grouping by modified nucleotides
}
