% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_storm.R
\name{storm_metricsScatterPlot_byPos}
\alias{storm_metricsScatterPlot_byPos}
\title{STORM scatterplot of scores by txcoor}
\usage{
storm_metricsScatterPlot_byPos(
  STORM,
  metrics = NULL,
  title = "",
  na.rm = FALSE
)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{metrics}{character. Selected metrics to plot. By default plots all metrics}

\item{title}{character. Title to be shown on top of plot}
}
\description{
Plot scatterplots of scores in STORM$RES by txcoor and colored by gene
}
