% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{storm_plot_confMat}
\alias{storm_plot_confMat}
\title{Confussion matrix plot}
\usage{
storm_plot_confMat(STORM, title = "", pointAlpha = 0.5)
}
\arguments{
\item{pointAlpha}{}
}
\description{
Confussion matrix plot
}
