% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{storm_STORM}
\alias{storm_STORM}
\title{STORM object constructor from META table}
\usage{
storm_STORM(META, genome = NULL, geneAnnot = NULL, nCores = 1, verb = FALSE)
}
\arguments{
\item{META}{data.frame Experimental design}

\item{verb}{}
}
\description{
STORM object constructor from META table
}
