% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-functions.R
\name{summarize_model}
\alias{summarize_model}
\title{Compute model summary.}
\usage{
summarize_model(model, label = NULL)
}
\arguments{
\item{model}{Fitted model object.}

\item{label}{Optional parameter to add a label to the resulting \code{data.table}.}
}
\value{
\code{data.table} object with coefficients and CIs.
}
\description{
Compute coefficients and confidence intervals from a fitted model object (e.g., from \code{lme4::lmer} or \code{limma::lmFit}).
}
\author{
Ayush Noori
}
