% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_BrTable}
\alias{test_BrTable}
\title{Test brain donor table}
\format{
A \code{data.frame} with two columns: \code{Sample} and \code{BrNum}.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/test_BrTable.R}
}
\usage{
test_BrTable
}
\description{
A test \code{data.frame} with example data for samples with brain donor ids (BrNums).
}
\keyword{datasets}
