% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_snps}
\alias{test_snps}
\title{TODO called}
\format{
TODO.

A \code{matrix} with zeros and ones representing the genotypes.
}
\source{
<https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/called.R
Test SNP matrix

A test SNP input \code{matrix}.

\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/test_snps.R}
}
\usage{
test_snps
}
\description{
TODO Description.
}
\keyword{datasets}
