% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{trainLogModel_allRNAmods}
\alias{trainLogModel_allRNAmods}
\title{Train all logistic models}
\usage{
trainLogModel_allRNAmods(
  STORM,
  thrRange = seq(-0.1, 0.5, 0.01),
  accNArate = TRUE,
  minMCC = 0.7
)
}
\arguments{
\item{STORM}{list. STORM object}

\item{thrRange}{numeric vector. set of thresholds used to search for a maximum
MatthewsCorrelationCoefficient}

\item{accNArate}{logical. Account for NA rate for model selection}

\item{minMCC}{numeric. Minimum MCC needed for RNAmod model to be output.
Set to -1 to force all models to be output.}
}
\description{
Train all logistic models
}
