% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{trainLogModel_RNAmods}
\alias{trainLogModel_RNAmods}
\title{Logistic Model training using STORM$CALLS object.}
\usage{
trainLogModel_RNAmods(
  STORM,
  isModFun,
  RNAmod,
  thresholds,
  accNArate = TRUE,
  minMCC = 0.7
)
}
\arguments{
\item{minMCC}{}
}
\description{
Logistic Model training using STORM$CALLS object.
}
