% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vList}
\alias{vList}
\title{Variables list for experimental design table creation}
\format{
list
}
\usage{
vList
}
\description{
Variables list for experimental design table creation
}
\keyword{datasets}
