% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx-functions.R
\name{write_excel}
\alias{write_excel}
\title{Write to Excel file.}
\usage{
write_excel(table, path, overwrite, ...)
}
\arguments{
\item{table}{Table to add to worksheet.}

\item{path}{Path to save Excel file.}

\item{overwrite}{If \code{TRUE}, overwrite any existing file.}

\item{...}{Other arguments to be passed to \code{brainstorm::add_worksheet()}.}
}
\description{
Combine \code{brainstorm} functions to write a single data table to an Excel file with a single worksheet. For finer control, functions can be called individually. Depends on the \code{openxlsx} package.
}
\author{
Ayush Noori
}
