% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_2OmeScore}
\alias{add_2OmeScore}
\title{Add 2-O-methylation score to STORM-seq object}
\usage{
add_2OmeScore(
  STORM,
  lib_LowdNTPs,
  lib_HighdNTPs,
  newColName = "auto",
  neighFlankSize = 5,
  perNuc_Zscaling = F,
  onNucs = c("A", "C", "G", "T"),
  minCov = 50
)
}
\arguments{
\item{minCov}{}
}
\value{

}
\description{
Add 2-O-methylation score to STORM-seq object
}
