% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_CtoT_MRD}
\alias{add_CtoT_MRD}
\title{Add C->T misincorporation difference (CtoT_MRD) to STORM object}
\usage{
add_CtoT_MRD(
  STORM,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("C"),
  minCov = 50
)
}
\arguments{
\item{onNucs}{}
}
\value{

}
\description{
Add C->T misincorporation difference (CtoT_MRD) to STORM object
}
