% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_CytPer}
\alias{add_CytPer}
\title{Add cytidine persistance index to STORM object}
\usage{
add_CytPer(STORM, group_A, group_B, newColName = "auto", onNucs = c("C"))
}
\arguments{
\item{onNucs}{}
}
\value{

}
\description{
Add cytidine persistance index to STORM object
}
