% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_default_metrics_v1}
\alias{add_default_metrics_v1}
\title{Add all default metrics for STORMseq v0.0.1}
\usage{
add_default_metrics_v1(STORM)
}
\arguments{
\item{STORM}{}
}
\value{

}
\description{
Add all default metrics for STORMseq v0.0.1
}
