% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_DRD}
\alias{add_DRD}
\title{Add deletion rate difference to STORM object}
\usage{
add_DRD(
  STORM,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T")
)
}
\arguments{
\item{onNucs}{}
}
\value{

}
\description{
Add deletion rate difference to STORM object
}
