% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_ER}
\alias{add_ER}
\title{Add read-end to coverage ratio}
\usage{
add_ER(
  STORM,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50
)
}
\arguments{
\item{minCov}{}
}
\value{

}
\description{
Add read-end to coverage ratio
}
