% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_MR}
\alias{add_MR}
\title{Add Missincorporation rate}
\usage{
add_MR(
  STORM,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50
)
}
\arguments{
\item{onNucs}{}
}
\value{

}
\description{
Add Missincorporation rate
}
