% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_scoreA3p}
\alias{add_scoreA3p}
\alias{storm_add_scoreA3p}
\title{Add ScoreA-3prime to a STORM object}
\usage{
add_scoreA3p(
  STORM,
  group_A,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minMedCov = 30
)
}
\arguments{
\item{minMedCov}{}
}
\value{

}
\description{
Add ScoreA-3prime to a STORM object
}
