% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{add_SRD}
\alias{add_SRD}
\title{Add Start Ratio Difference}
\usage{
add_SRD(
  STORM,
  group_A,
  group_B,
  newColName = "auto",
  onNucs = c("A", "C", "G", "T"),
  minCov = 50
)
}
\arguments{
\item{minCov}{}
}
\value{

}
\description{
Calculates the start ratio difference of group_A minus that
of group_B per set of samples in a STORM object.
}
