% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{alignSTAR}
\alias{alignSTAR}
\title{STAR alignment}
\usage{
alignSTAR(
  read1Files,
  STARgenomeDir,
  pairedEnd = TRUE,
  zipped = TRUE,
  nCores = 4,
  alignEndsType = "Local",
  alignIntronMax = 0,
  outFilterMultimapNmax = 10,
  outDir,
  otherSTARparams = "",
  dry = FALSE,
  tmpDir = NULL,
  logSumm = TRUE
)
}
\arguments{
\item{read1Files}{character. Path to R1 FASTQ files}

\item{STARgenomeDir}{character. Path to STAR genome to map to}

\item{pairedEnd}{logical. Indicating if to expect a Read2 FASTQ file
File will be automatically looked for but both need to include "_R1", and "_R2"
in their respective file names.}

\item{zipped}{logical. TRUE as default for zipped FASTQ files, will be read with zcat.
If set to FALSE FASTQ files will be read with cat, as not zipped.}

\item{nCores}{numeric. Number of cores used for alignment and sorting processes.}

\item{alignEndsType}{character. type of read ends alignment (See STAR manual for more info)}

\item{alignIntronMax}{numeric. maximum intron size, if 0, max intron size will be determined
by default as (2ˆwinBinNbits)*winAnchorDistNbins (See STAR manual for more info)}

\item{outFilterMultimapNmax}{numeric. Threshold for which a read will have multiple
mappings. Default is 10. For unique alignment change the value to 1.}

\item{outDir}{character. Path to output directory}

\item{otherSTARparams}{character. Additional parameters not covered by the arguments
of this function can be added here, separated by spaces.}

\item{dry}{logical. If set to TRUE the alignment is not performed, only output
are the paths to expected output files.}

\item{tmpDir}{character. Path to directory to be used to generate
intermediate files.}

\item{logSumm}{logical. Set to FALSE so final logs are not appended to
\emph{mappingSummary.txt}.}
}
\description{
Wrapper to perform alignment of sequencing reads to a reference genome
using STAR (Dobin) and sorting and indexing using Samtools.
}
