% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{alignSTAR}
\alias{alignSTAR}
\title{STAR alignment}
\usage{
alignSTAR(
  read1Files,
  STARgenomeDir,
  pairedEnd = TRUE,
  zipped = TRUE,
  nCores = 4,
  alignEndsType = "Local",
  outSAMtype = "BAM Unsorted",
  alignIntronMax = 0,
  outFilterMultimapNmax = 10,
  outDir
)
}
\arguments{
\item{read1Files}{character. Path to R1 FASTQ files}

\item{STARgenomeDir}{character. Path to STAR genome to map to}

\item{alignIntronMax}{}
}
\value{

}
\description{
Wrapper to perform alignment of sequencing reads to a reference genome
using STAR (Dobin) and sorting with Samtools.
}
