% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainstorm-package.R
\docType{package}
\name{brainstorm-package}
\alias{brainstorm-package}
\alias{brainstorm}
\title{Personal utility functions.}
\description{
This package houses personal utility functions which I reuse in my R scripts.
}
\references{
\url{https://www.ayushnoori.com/brainstorm}
}
\author{
Ayush Noori \email{anoori1@mgh.harvard.edu}
}
