% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainstorm-package.R
\docType{package}
\name{brainstorm-package}
\alias{brainstorm-package}
\alias{brainstorm}
\title{Personal utility functions.}
\description{
This package houses personal utility functions which I reuse in my R scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.ayushnoori.com/brainstorm}
  \item \url{https://github.com/ayushnoori/brainstorm}
  \item Report bugs at \url{https://github.com/ayushnoori/brainstorm/issues}
}

}
\author{
\strong{Maintainer}: Ayush Noori \email{anoori1@mgh.harvard.edu} (\href{https://orcid.org/0000-0003-1420-1236}{ORCID})

}
\keyword{internal}
