% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_called.R
\name{filter_called}
\alias{filter_called}
\title{Title}
\usage{
filter_called(
  called,
  DP_min = 5,
  DP_max = 80,
  ref_call = 1,
  alt_call = 1,
  overlaps = 1,
  min_VDB = 0.1
)
}
\arguments{
\item{called}{CollapsedVCF containing filtered snp info for rna data}

\item{min_VDB}{}
}
\value{

}
\description{
Title
}
\examples{
geno_filtered <- filter_called(genotyped)
dim(called)

#222 5568
called_filter <- filter_called(called)
dim(called_filter)
#748 5568
}
