% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hlp_assign_scores}
\alias{hlp_assign_scores}
\title{Assign scores to respective RNAmods in STORM$CALLS}
\usage{
hlp_assign_scores(STORM, RNAmod, scores)
}
\arguments{
\item{STORM}{list. STORM object as output by \code{\link{storm_STORM}}}

\item{scores}{}
}
\value{

}
\description{
Assign scores to respective RNAmods in STORM$CALLS
}
