% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{libComplexReport}
\alias{libComplexReport}
\title{Library complexity report}
\usage{
libComplexReport(
  META,
  maxExtrapolation = 2010000,
  steps = 1e+05,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\description{
Library complexity report
}
