% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snpsGeno.R
\name{make_snpsGeno}
\alias{make_snpsGeno}
\title{Make the snp array for genotype data}
\usage{
make_snpsGeno(genotyped)
}
\arguments{
\item{genotyped}{object from variantAnnotation::readVcf}
}
\value{
snp array encoded as integers
}
\description{
Make the snp array for genotype data
}
\examples{
genotyped <- readVcf("Data/LIBD_Brain_merged_topmed_RNASeq_annotated_variants_043020.vcf", "hg38")
snpsGeno <- make_snpsGeno(genotyped)
snpsGeno[1:5, 1:5]
head(colnames(snpsGeno))
}
