% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snpsGeno.R
\name{make_snpsGeno}
\alias{make_snpsGeno}
\title{Make the snp array for genotype data}
\usage{
make_snpsGeno(genotyped)
}
\arguments{
\item{genotyped}{object from variantAnnotation::readVcf}
}
\value{
snp array encoded as integers
}
\description{
Make the snp array for genotype data
}
\examples{
snpsGeno <- make_snpsGeno(snpsGeno_VCF)
snpsGeno[1:5, 1:5]
}
