% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{mkBedFromFastaTxOme}
\alias{mkBedFromFastaTxOme}
\title{Make BED from FASTA Transcriptome}
\usage{
mkBedFromFastaTxOme(fastaTxOme, outFile = "auto")
}
\arguments{
\item{outFile}{}
}
\description{
Make BED Gene Annotation from a FASTA genome. Considering that the sequences
in the FASTA file represent intron-less genes.
}
