% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_in_WEXAC.R
\name{mkTranscriptome}
\alias{mkTranscriptome}
\title{Make transcriptome}
\usage{
mkTranscriptome(fastaGenome, bedAnnotation, outFile = "auto", nCores)
}
\arguments{
\item{nCores}{}
}
\description{
Extract transcriptome sequences from genome and generate FASTA file
}
