% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown-functions.R
\name{show_table}
\alias{show_table}
\title{Show Markdown table.}
\usage{
show_table(table, height = "300px")
}
\arguments{
\item{table}{Table to display.}

\item{height}{Optional parameter to set height of scroll box. Default is 300px.}
}
\value{
\code{kable} table object.
}
\description{
Read table in Markdown document. Depends on \code{knitr} and \code{kableExtra} packages.
}
\author{
Ayush Noori
}
