% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_STORM.R
\name{storm_META}
\alias{storm_META}
\title{Create STORM META - Experiment design matrix}
\usage{
storm_META(fileNames, varsList, groupVars, setVars, idVars, outDir)
}
\arguments{
\item{outDir}{}
}
\value{

}
\description{
STORM META constructor from RDS file names
}
