% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_STORM.R
\name{storm_STORM}
\alias{storm_STORM}
\title{STORM object constructor from META table}
\usage{
storm_STORM(META, genome = NULL, geneAnnot = NULL, nCores = 1)
}
\arguments{
\item{nCores}{}
}
\value{

}
\description{
STORM object constructor from META table
}
