% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_snps}
\alias{test_snps}
\title{Test SNP matrix}
\format{
A \code{matrix} with zeros and ones representing the genotypes.
}
\source{
\url{https://github.com/LieberInstitute/brainstorm/blob/master/data-raw/test_snps.R}
}
\usage{
test_snps
}
\description{
A test SNP input \code{matrix}.
}
\keyword{datasets}
